Modularize the Event Header
